Build: Makefile: do not pass BUILD_KEYS in release build
Also replace build_desc with build ID in eng/userdebug build.

Starting with Oreo, individual builds are identified with a new build ID format, in the form of PVBB.YYMMDD.HHMMSS
The P part represents the first letter of the code name of the platform release, e.g. O is Oreo.
The V part represents a supported vertical. By convention, P represents the primary platform branch.
The BB part represents a alpha numeric code which allows Google to identify the exact code branch that the build was made from.
The YYMMDD part identifies the exact date at which a build was made.
The HHMMSS part identifies the exact time at which a build was made.